WIP: Do not merge. Rework how-to/store section.
I moved the information on sector sizes to the start of the
section and included more links.

I'm going to turn the "Large Files" page into a mini-tutorial
on how to split data.

I am planning to create a separate "offline deals" page, as
well as include some more IPLD/Dumbo Drop techniques in an
advanced techniques section somewhere.