Remove upnp section (#236)
Due to the fact that the port number of the miner will be put on chain by the `set-addrs` call, it is ill-advised to use UPnP as it will potentially provide a new port if the process is restarted. This was my mistake as I suggested this change before I knew about putting the address on chain.