Mining architectures. HW requirements revamp. (#450)
This PR increases the context around setting up a Filecoin mining operation.

The hardware requirements page has been reworked, explaining a bit better how
the requirements apply and providing not only general but also specific
requirements for each of the mining operations that need to be performed
(sealing phases etc).

A new "mining architecture" guide is added, with the PL v3 mining setup added
in it. It should hopefully include additional architectures from other parties
in the future.