Update installation.md
Adding in workarounds for folks with issues running `sudo make install`. Not sure if this is necessary / will likely be fixed in the future. Let me know if you don't want to add this to the doc. I shared the workaround in this issue as well: https://github.com/filecoin-project/lotus/issues/3076