Move gas documentation to how-filecoin-works
Documentation on gas and fees was in the message pool guide, but this is not
information that is specific to lotus, it actually belongs to the network.

This moves information and definitions about gas and fees to the
how-filecoin-works guide. It takes in some improvements from the Gas Fees blog
post and adds some more.