Re-adds deleted line.
In PR #990, I deleted what I thought was a duplicate line. However, it turns out that it's not a duplicate command, but it does look very similar. Having both commands on separate lines might cause users, like myself, to assume it's a duplicate and just run the top line. Adding the 2nd part of the command with `&&` should mitigate this potential issue.